Time: 0 ms (100%), Space: 7.6 MB (99.29%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPerfectSquare(int num) 
     {
-        double k = sqrt(num);
-        return num == pow(int(k), 2);    
+        int k = sqrt(num);
+        return num == pow(k, 2);    
     }
 };
